unstop/requirements.txt
PyMuPDF==1.23.8
docx2txt==0.8
google-generativeai==0.3.2
Pillow==10.2.0
gradio==4.19.2
python-dotenv==1.0.1
markupsafe==2.1.5 